+++ boofuzz/sessions.py
@@ -363,6 +363,8 @@ class Session(pgraph.Graph):
                                             Default None.
         post_test_case_callbacks (list of method): The registered method will be called after each fuzz test case.
                                                   Default None.
+        post_start_target_callbacks (list of method): Method(s) will be called after the target is started or restarted,
+                                                      say, by a process monitor.
         web_port (int):         Port for monitoring fuzzing campaign via a web browser. Default 26000.
         keep_web_open (bool):     Keep the webinterface open after session completion. Default True.
         fuzz_loggers (list of ifuzz_logger.IFuzzLogger): For saving test data and results.. Default Log to STDOUT.
@@ -408,6 +410,7 @@ class Session(pgraph.Graph):
         restart_timeout=None,
         pre_send_callbacks=None,
         post_test_case_callbacks=None,
+        post_start_target_callbacks=None,
         fuzz_loggers=None,
         fuzz_db_keep_only_n_pass_cases=0,
         receive_data_after_each_request=True,
@@ -478,13 +481,21 @@ class Session(pgraph.Graph):
         else:
             post_test_case_methods = post_test_case_callbacks
 
+        if post_start_target_callbacks is None:
+            post_start_target_methods = []
+        else:
+            post_start_target_methods = post_start_target_callbacks
+
         if restart_callbacks is None:
             restart_methods = []
         else:
             restart_methods = restart_callbacks
 
         self._callback_monitor = CallbackMonitor(
-            on_pre_send=pre_send_methods, on_post_send=post_test_case_methods, on_restart_target=restart_methods
+            on_pre_send=pre_send_methods,
+            on_post_send=post_test_case_methods,
+            on_restart_target=restart_methods,
+            on_post_start_target=post_start_target_methods,
         )
 
         self.total_num_mutations = 0
@@ -771,8 +782,7 @@ class Session(pgraph.Graph):
             self.server_init()
 
         try:
-            for monitor in self.targets[0].monitors:
-                monitor.start_target()
+            self._start_target(self.targets[0])
 
             if self._reuse_target_connection:
                 self.targets[0].open()
@@ -822,6 +832,16 @@ class Session(pgraph.Graph):
             raise
         finally:
             self._fuzz_data_logger.close_test()
+
+    def _start_target(self, target):
+        started = False
+        for monitor in target.monitors:
+            if monitor.start_target():
+                started = True
+                break
+        if started:
+            for monitor in target.monitors:
+                monitor.post_start_target(target=target, fuzz_data_logger=self._fuzz_data_logger, session=self)
 
     def import_file(self):
         """
@@ -1093,18 +1113,19 @@ class Session(pgraph.Graph):
         #       a custom callback. wtf?
 
         self._fuzz_data_logger.open_test_step("Restarting target")
+        restarted = False
         if len(self.on_failure) > 0:
             for f in self.on_failure:
                 self._fuzz_data_logger.open_test_step("Calling registered on_failure method")
                 f(logger=self._fuzz_data_logger)
+            restarted = True
         # vm restarting is the preferred method so try that before monitors.
         elif target.vmcontrol:
             self._fuzz_data_logger.log_info("Restarting target virtual machine")
             target.vmcontrol.restart_target()
-
+            restarted = True
         # we always have at least one monitor; a Callback Monitor that handles all callbacks.
         else:
-            restarted = False
             for monitor in target.monitors:
                 self._fuzz_data_logger.log_info("Restarting target process using {}".format(monitor.__class__.__name__))
                 if monitor.restart_target(target=target, fuzz_data_logger=self._fuzz_data_logger, session=self):
@@ -1114,12 +1135,14 @@ class Session(pgraph.Graph):
                     restarted = True
                     break
 
-            # no monitor can restart
-            if not restarted:
-                self._fuzz_data_logger.log_info(
-                    "No reset handler available... sleeping for {} seconds".format(self.restart_sleep_time)
-                )
-                time.sleep(self.restart_sleep_time)
+        if restarted:
+            for monitor in target.monitors:
+                monitor.post_start_target(target=self.targets[0], fuzz_data_logger=self._fuzz_data_logger, session=self)
+        else:
+            self._fuzz_data_logger.log_info(
+                "No reset handler available... sleeping for {} seconds".format(self.restart_sleep_time)
+            )
+            time.sleep(self.restart_sleep_time)
 
         # pass specified target parameters to the PED-RPC server to re-establish connections.
         target.monitors_alive()